Add database and file dialog functionality, refactor interface
Code is restructured to better accommodate components by creating different packages for MDI area and top tab widgets. The interface file is accordingly moved. New functionality added enables the opening and reloading of databases as well as calling file and directory dialogs. The changes also reflect in the updates made to the canguru application entry point, importing the correct modules.